pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "navvy"
version = "1.0.0"
description = "Automation tool for managing Git repositories with AI (OpenAI)."
readme = "README.md"
authors = [{ name = "Rofly António", email = "mitangerofly@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["automation", "ai", "openai", "git", "code"]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "certifi==2024.8.30",
    "distro==1.9.0",
    "gitdb==4.0.11",
    "GitPython==3.1.43",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.27.2",
    "idna==3.10",
    "jiter==0.7.1",
    "openai==1.54.5",
    "pydantic==2.9.2",
    "pydantic_core==2.23.4",
    "smmap==5.0.1",
    "sniffio==1.3.1",
    "tqdm==4.67.0",
    "typing_extensions==4.12.2"
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/itsrofly/navvy-package"
